Update students page with links to personal websites
This commit updates the students page on the theory group website. It adds links to each student's personal websites, if available. It also fixes the ordering of students names to follow alphabetical order by last name. It also changes students' names to their preferred names, which is usually the name they use when publishing papers.