Update postdoc alumni page
This commit updates the postdoc alumni page. It uses "https" instead of "http"
wherever applicable. It also updates the people's job information to our best
knowledge. For example, it updates people from assistant professor to professor.
It also tries to update expired links with the most up to date link.