Use https and update links for home page, faculty page, and positions page
This commit mainly replaces the 'http' link with 'https'. Why? It
is because some browsers automatically blocks all the 'http' pages
since they are not considered secure. This will prevent potential
students and other potential researchers to know more about the
theory group here at UofT. It also updates the links in the home
page and posiitions to be the newest links. This will help reducing
loading speed and provide a better experience for users to browse
theory group's website.