Fix formatting on alumni and postdocs page, add Malcolm on faculty page
This commit adds indentation to the list of alumni and list of postdoc
alumni to make it look better. It also adds Malcolm to the faculty
and staff page as an IT support specialist. It also adds a link to
Nathan's personal website. Most of the code changes are fixing the
code indentation on the faculty page but nothing is changed for
faculty info.