Reformat admin staff on faculty page and unify indentation on html code
This commit reformats the administrative part on faculty page as the
previous design does not look good. It also tries to unify the
indentation on all html pages to be 2 spaces. This will leave the
future contributors a cleaner code base to work with. It also deletes
the grad.html page which does not belong to UofT.