Fix font loading bug
When we open the webpage and inspect on the webpage, the following
error message pops up in the console: Blocked loading mixed active
content
"http://fonts.googleapis.com/css?family=Yanone+Kaffeesatz%7CDroid+Sans"
This is due to we have upgrade the whole website to https but the
font loading url is using http thus the browser is blocking it since
it wants to load from all https resources. I have checked that the
https version of the googleapi is available thus I am changing it
to https to make the font css work correctly as coded.