Fix the padding issue in the slider in the home page
The home page has a weird padding issue where there are some empty
space to the right but no empty space to the left. The best solution
I have so far is to resize the images to be 960 by 300 pixels such
that they fill the entire window. This commit does that.